LPS-92553 Use new implementation for info display contributor and fields
