+++ modules/apps/adaptive-media/adaptive-media-blogs-web/src/main/java/com/liferay/adaptive/media/blogs/web/internal/asset/display/contributor/AMBlogsEntryContentInfoDisplayContributorField.java
@@ -12,14 +12,13 @@
  * details.
  */
 
-package com.liferay.document.library.web.internal.asset.display.contributor;
+package com.liferay.adaptive.media.blogs.web.internal.asset.display.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
-import com.liferay.document.library.util.DLURLHelper;
-import com.liferay.petra.string.StringPool;
+import com.liferay.adaptive.media.content.transformer.ContentTransformerHandler;
+import com.liferay.adaptive.media.content.transformer.constants.ContentTransformerContentTypes;
+import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.info.display.contributor.InfoDisplayContributorField;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
 import java.util.Locale;
 
@@ -30,21 +29,23 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alejandro Tardín
  */
 @Component(
-	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
-	service = AssetDisplayContributorField.class
+	property = {
+		"model.class.name=com.liferay.blogs.model.BlogsEntry",
+		"service.ranking:Integer=2"
+	},
+	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryDownloadURLAssetDisplayContributorField
-	implements AssetDisplayContributorField<FileEntry> {
+public class AMBlogsEntryContentInfoDisplayContributorField
+	implements InfoDisplayContributorField<BlogsEntry> {
 
 	@Override
 	public String getKey() {
-		return "downloadURL";
+		return "content";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "download-url");
+		return LanguageUtil.get(locale, "content");
 	}
 
 	@Override
@@ -53,17 +54,12 @@ public class DLFileEntryDownloadURLAssetDisplayContributorField
 	}
 
 	@Override
-	public String getValue(FileEntry fileEntry, Locale locale) {
-		try {
-			return _dlurlHelper.getDownloadURL(
-				fileEntry, fileEntry.getFileVersion(), null, StringPool.BLANK);
-		}
-		catch (Exception pe) {
-			return null;
-		}
+	public String getValue(BlogsEntry blogsEntry, Locale locale) {
+		return _contentTransformerHandler.transform(
+			ContentTransformerContentTypes.HTML, blogsEntry.getContent());
 	}
 
 	@Reference
-	private DLURLHelper _dlurlHelper;
+	private ContentTransformerHandler _contentTransformerHandler;
 
 }
