+++ modules/apps/blogs/blogs-reading-time/build.gradle
@@ -2,7 +2,7 @@ dependencies {
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	compileOnly project(":apps:asset:asset-display-api")
 	compileOnly project(":apps:blogs:blogs-api")
+	compileOnly project(":apps:info:info-api")
 	compileOnly project(":apps:reading-time:reading-time-api")
 }
