+++ modules/apps/blogs/blogs-reading-time/src/main/java/com/liferay/blogs/reading/time/internal/asset/display/contributor/BlogsEntryReadingTimeInfoDisplayContributorField.java
@@ -12,31 +12,35 @@
  * details.
  */
 
-package com.liferay.blogs.web.internal.asset.display.contributor;
+package com.liferay.blogs.reading.time.internal.asset.display.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
 import com.liferay.blogs.model.BlogsEntry;
+import com.liferay.info.display.contributor.InfoDisplayContributorField;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
+import com.liferay.reading.time.message.ReadingTimeMessageProvider;
+import com.liferay.reading.time.model.ReadingTimeEntry;
+import com.liferay.reading.time.service.ReadingTimeEntryLocalService;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Alejandro Tardín
  */
 @Component(
 	property = "model.class.name=com.liferay.blogs.model.BlogsEntry",
-	service = AssetDisplayContributorField.class
+	service = InfoDisplayContributorField.class
 )
-public class BlogsEntryCoverImageCaptionAssetDisplayContributorField
-	implements AssetDisplayContributorField<BlogsEntry> {
+public class BlogsEntryReadingTimeInfoDisplayContributorField
+	implements InfoDisplayContributorField<BlogsEntry> {
 
 	@Override
 	public String getKey() {
-		return "coverImageCaption";
+		return "readingTime";
 	}
 
 	@Override
@@ -44,7 +48,7 @@ public class BlogsEntryCoverImageCaptionAssetDisplayContributorField
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			locale, getClass());
 
-		return LanguageUtil.get(resourceBundle, "cover-image-caption");
+		return LanguageUtil.get(resourceBundle, "reading-time");
 	}
 
 	@Override
@@ -54,7 +58,17 @@ public class BlogsEntryCoverImageCaptionAssetDisplayContributorField
 
 	@Override
 	public String getValue(BlogsEntry blogsEntry, Locale locale) {
-		return blogsEntry.getCoverImageCaption();
+		ReadingTimeEntry readingTimeEntry =
+			_readingTimeEntryLocalService.fetchOrAddReadingTimeEntry(
+				blogsEntry);
+
+		return _readingTimeMessageProvider.provide(readingTimeEntry, locale);
 	}
+
+	@Reference
+	private ReadingTimeEntryLocalService _readingTimeEntryLocalService;
+
+	@Reference
+	private ReadingTimeMessageProvider _readingTimeMessageProvider;
 
 }
