+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/asset/display/contributor/DLFileEntryPreviewImageInfoDisplayContributorField.java
@@ -14,7 +14,8 @@
 
 package com.liferay.document.library.web.internal.asset.display.contributor;
 
-import com.liferay.asset.display.contributor.AssetDisplayContributorField;
+import com.liferay.document.library.util.DLURLHelper;
+import com.liferay.info.display.contributor.InfoDisplayContributorField;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
@@ -22,36 +23,45 @@ import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Alejandro Tardín
  */
 @Component(
 	property = "model.class.name=com.liferay.document.library.kernel.model.DLFileEntry",
-	service = AssetDisplayContributorField.class
+	service = InfoDisplayContributorField.class
 )
-public class DLFileEntryFileNameAssetDisplayContributorField
-	implements AssetDisplayContributorField<FileEntry> {
+public class DLFileEntryPreviewImageInfoDisplayContributorField
+	implements InfoDisplayContributorField<FileEntry> {
 
 	@Override
 	public String getKey() {
-		return "fileName";
+		return "previewImage";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
 		return LanguageUtil.get(
-			ResourceBundleUtil.getBundle(locale, getClass()), "file-name");
+			ResourceBundleUtil.getBundle(locale, getClass()), "preview-image");
 	}
 
 	@Override
 	public String getType() {
-		return "text";
+		return "image";
 	}
 
 	@Override
 	public String getValue(FileEntry fileEntry, Locale locale) {
-		return fileEntry.getFileName();
+		try {
+			return _dlurlHelper.getImagePreviewURL(fileEntry, null);
+		}
+		catch (Exception pe) {
+			return null;
+		}
 	}
+
+	@Reference
+	private DLURLHelper _dlurlHelper;
 
 }
